ACLs: Extract AuthService implementation
